Rename /global-scopes to /scopes
