+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/ScopesApiService.java
@@ -9,6 +9,9 @@ import org.apache.cxf.jaxrs.ext.multipart.Multipart;
 
 import org.wso2.carbon.apimgt.api.APIManagementException;
 
+import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.ErrorDTO;
+import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.ScopeDTO;
+import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.ScopeListDTO;
 
 import java.util.List;
 
@@ -19,5 +22,10 @@ import javax.ws.rs.core.SecurityContext;
 
 
 public interface ScopesApiService {
+      public Response addGlobalScope(ScopeDTO body, MessageContext messageContext) throws APIManagementException;
+      public Response deleteGlobalScope(String scopeId, MessageContext messageContext) throws APIManagementException;
+      public Response getGlobalScope(String scopeId, MessageContext messageContext) throws APIManagementException;
+      public Response getGlobalScopes(MessageContext messageContext) throws APIManagementException;
+      public Response updateGlobalScope(String scopeId, ScopeDTO body, MessageContext messageContext) throws APIManagementException;
       public Response validateScope(String name, MessageContext messageContext) throws APIManagementException;
 }
