+++ docs/apidocs/publisher/v1/index.html
@@ -295,17 +295,6 @@
                             </ul>
                         </li>
                         <hr/>
-                        <span class="class-name">Global Scopes</span>
-                        <li class="section-box">
-                            <ul class="nav nav-stacked">
-                                        <li id="op-addGlobalScope"><a href="#!/operations#GlobalScopes#addGlobalScope">Add a new Global Scope</a></li>
-                                        <li id="op-deleteGlobalScope"><a href="#!/operations#GlobalScopes#deleteGlobalScope">Delete a Global Scope</a></li>
-                                        <li id="op-getGlobalScope"><a href="#!/operations#GlobalScopes#getGlobalScope">Get a Global Scope by Scope Id</a></li>
-                                        <li id="op-getGlobalScopes"><a href="#!/operations#GlobalScopes#getGlobalScopes">Get all available Global Scopes</a></li>
-                                        <li id="op-updateGlobalScope"><a href="#!/operations#GlobalScopes#updateGlobalScope">Update a Global Scope</a></li>
-                            </ul>
-                        </li>
-                        <hr/>
                         <span class="class-name">GraphQL Schema</span>
                         <li class="section-box">
                             <ul class="nav nav-stacked">
@@ -346,6 +335,17 @@
                         <li class="section-box">
                             <ul class="nav nav-stacked">
                                         <li id="op-validateScope"><a href="#!/operations#Scope#validateScope">Check given scope name is already exist</a></li>
+                            </ul>
+                        </li>
+                        <hr/>
+                        <span class="class-name">Scopes</span>
+                        <li class="section-box">
+                            <ul class="nav nav-stacked">
+                                        <li id="op-addGlobalScope"><a href="#!/operations#Scopes#addGlobalScope">Add a new Global Scope</a></li>
+                                        <li id="op-deleteGlobalScope"><a href="#!/operations#Scopes#deleteGlobalScope">Delete a Global Scope</a></li>
+                                        <li id="op-getGlobalScope"><a href="#!/operations#Scopes#getGlobalScope">Get a Global Scope by Scope Id</a></li>
+                                        <li id="op-getGlobalScopes"><a href="#!/operations#Scopes#getGlobalScopes">Get all available Global Scopes</a></li>
+                                        <li id="op-updateGlobalScope"><a href="#!/operations#Scopes#updateGlobalScope">Update a Global Scope</a></li>
                             </ul>
                         </li>
                         <hr/>
@@ -703,11 +703,6 @@
                         </li>
                         <li class="section-box">
                             <ul class="nav nav-stacked">
-                                <li id="def-GlobalScopeList"><a href="#!/models#GlobalScopeList">GlobalScopeList</a></li>
-                            </ul>
-                        </li>
-                        <li class="section-box">
-                            <ul class="nav nav-stacked">
                                 <li id="def-GraphQLSchema"><a href="#!/models#GraphQLSchema">GraphQLSchema</a></li>
                             </ul>
                         </li>
@@ -834,6 +829,11 @@
                         <li class="section-box">
                             <ul class="nav nav-stacked">
                                 <li id="def-Scope"><a href="#!/models#Scope">Scope</a></li>
+                            </ul>
+                        </li>
+                        <li class="section-box">
+                            <ul class="nav nav-stacked">
+                                <li id="def-ScopeList"><a href="#!/models#ScopeList">ScopeList</a></li>
                             </ul>
                         </li>
                         <li class="section-box">
