--- docs/apidocs/publisher/v1/models/GlobalScopeList.html
@@ -1,54 +1,0 @@
-<div class="main-content">
-
-        <h2>GlobalScopeList</h2>
-            <table class="table table-hover">
-                <colgroup>
-                    <col style="width: 10%;"/>
-                    <col style="width: 30%;"/>
-                    <col style="width: 10%;"/>
-                    <col style="width: 20%;"/>
-                    <col style="width: 30%;"/>
-                </colgroup>
-                <thead>
-                <tr>
-                    <th>Name</th>
-                    <th>Description</th>
-                    <th>Schema</th>
-                    <th>Default</th>
-                    <th>Example</th>
-                </tr>
-                </thead>
-                <tbody>
-                    <tr>
-                        <td class="parameter">
-                            <p>count</p>
-                            <p class="param-required">
-                                
-                                optional
-                            </p>
-                        </td>
-                        <td class="parameter"><p class="marked">Number of Global Scopes returned. </p></td>
-                        <td class="parameter">
-                                <a href="#!/models#integer">integer</a>
-                        </td>
-                        <td class="parameter"> <p class="marked">null</p></td>
-                        <td class="parameter"> <p class="marked">1</p></td>
-                    </tr>
-                    <tr>
-                        <td class="parameter">
-                            <p>list</p>
-                            <p class="param-required">
-                                
-                                optional
-                            </p>
-                        </td>
-                        <td class="parameter"><p class="marked"></p></td>
-                        <td class="parameter">
-                                <a href="#!/models#array">array</a>
-                        </td>
-                        <td class="parameter"> <p class="marked">null</p></td>
-                        <td class="parameter"> <p class="marked"></p></td>
-                    </tr>
-                </tbody>
-            </table>
-</div>
